requirements.txt
boto==2.49.0
boto3==1.9.188
botocore==1.12.188
certifi==2019.6.16
chardet==3.0.4
Django==2.1.7
djangorestframework==3.9.4
django-cors-headers==2.4.0
docutils==0.14
idna==2.8
jmespath==0.9.4
nltk==3.4.4
numpy==1.16.4
pandas==0.24.2
phonenumbers==8.10.14
Pillow==6.1.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
s3fs==0.2.2
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
image==1.5.27
pytesseract==0.2.7
spacy==2.1.4